requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
discord.py==2.3.2
frozenlist==1.4.1
idna==3.6
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
peewee==3.17.1
Pygments==2.17.2
python-dotenv==1.0.1
rich==13.7.1
yarl==1.9.4
